Final edits, custom hooks improved
